Introduce different data types for the different kinds of variables
So now we have `Shape.var', `Region.t' and `EffectVar.t'; for
convenience, we still keep `Var.t' as the sum of the three. Each
of these variable types may represent a bound (polymorphic) or a
meta variable of its kind, and substitution is possible by
_zonking_.

Thus, this patch brings a bit more of type-safety, and an easy
way to substitute meta variables by polymorphic ones.

#refactor #cleanup