Constrain the result shape of logical operators to `Shape.Bot'
Logical operators are defined by the C standard to return either
`0' (false) or `1' (true). They cannot be sensibly converted into
any other shape than _|_ (bottom). We must allow such a conversion,
but it must be marked as unsafe. However, if we keep the result
shape of these operators polymoprhic, it will always be considered
safe to instantiate it with any arbitrary shape.