Build an effect-abstraction of the program during inference
During effect inference we map variables to shape schemes and
locations to effects. We do not construct a new AST annotated
with effects, but we use a plain-old dictionary to store this
information. The CIL AST plus this dictionary constitutes an
effect-based abstraction of the input C program.

Note that since we map effects to locations we do not track
the effects of individual CIL instructions (all of them are
mapped to the same location), but just the effects of the
original C statement.