Add a simple reachability checker and an uninitialized-read checker
The reachability engine can model check a function's CFG against
queries of the form:

  ... when P
  Q

Which, if I remember correctly, corresponds to `P EU Q' in CTL. Here
`P' and `Q' are predicates over computational effects.

The uninitialized-read checker runs a query `!writes EU reads' for
each of the local variables in a function's definition.