Fix typing of `StartOf' expressions
This was violating an invariant and thus causing a crash:
`of_lval' must return a `ref' shape, while `of_exp' must
not return a `ref' shape.

The `StartOf' operator simply obtains a pointer to the
region where the elements of the array are stored. In our
case this is equivalent to the `Lval' rule.

#bugfix #63a9b235d30e