Introduce `uninit' effects for tracking uninitialized memory regions
By treating uninitialization like an effect, we will be able to encode
read-before-initialization checkers the same way as the others: i.e.
as a pattern over effects, not having to look at variable declarations.
It is also useful to know whether a function may return a pointer to
uninitialized data. An effect `uninit(r)' can be interpreted as setting
`r' to some random unpredictable value.