Accept function types with unspecified arguments
The case I've seen so far is the use of function pointers with no
arguments specified. Naturally, when these function pointers are
instantiated with concrete function types, we may get unification
errors since there are no arguments to unify against. This is not
really a fatal error anyways, we can deal with unsoundness.