Refine model-checker's infinite loop avoidance
The previous approach of avoiding visiting the same edge twice is too
restrictive. For example in

    :loop
    S1
    if (c1) goto loop;
    S2

we would like to explore the path where the `if' is took once, resulting
in S1 being executed twice. But that cannot be done without re-visiting
the edge from S1 to the `if' statement. Yet taking this edge twice is
not a problem, as it is a sequential step. What we should avoid is to
bindly take the `goto'!

Now we count how many times and edge has been taken, which is more
flexible ---making easier to change the iteration bound. Further, we
count edges that jump to a labeled node only, since those are the ones
that count.