Assign a different region to each field of a structure shape
Until now all fields of a struct shared the same region. This hurts precision
and makes difficult to track operations on struct fields. For instance, given
a struct with two lock object fields, the double lock checker could not easily
distinguish which concrete lock was acquired or released. Also, the aliasing
of two fields of two different struct objects, would cause the two structs and
all their fields to be considered aliased.

Having each field its own region is a big win in terms of precision. However,
now it is a bit more cumbersome to detect object updates. For instance, if
we acquire a `s->lock', and later we overwrite `s', `s->lock' will probably
represent a different lock. However, we will notice this if we only observe
writes to the storage region of `s->lock', we need to observe updates to `s'
itself.

#precision #aliasing #structs