Introduce partial axioms to refine a skeleton of a function's shape scheme
It's now possible to take the template shape of a function (as computed by
`Shape.of_typ') and refine it by adding effects or even constraining variables.
Crucially, this makes very easy to provide axioms for functions taking structs
as parameters, without having to specify the exact shape of the struct object.

As a result the `Axioms' module became a mess and needed some refactoring. Now
axioms are split into: blind, partial, and onsite axioms ---see source docs.

#axioms #feature #refactoring