lenv: Evaluate rather than just looking up fact
When "learning" from an assignment, we can do better than just looking
up the RHS in the fact environemnt, we can try evaluating the RHS!
E.g., the value of `(! (! x))` may not be in our fact environment but,
if we know the value of `x`, then we can evaluate that expression.

Initially we didn't pay much attention to assignments, but note that

    if (f(x)) ...

will become

    tmp = f(x);
    if (tmp) ...

in CIL.

Linux makes extensive use of `__builtin_expect`, so this transformation
affects plenty of if-statements.